Isolate app layer from platform layer. You can now use strings/file/paths etc without pulling in the app layer by defining OC_NO_APP_LAYER before including orca.c
