+++ src/orca.c
@@ -25,6 +25,7 @@
     #include "platform/native_debug.c"
     #include "platform/unix_memory.c"
     #include "platform/osx_clock.c"
+    #include "platform/osx_path.c"
     #include "platform/posix_io.c"
     #include "platform/posix_thread.c"
     #include "platform/osx_platform.c"
@@ -39,6 +40,7 @@
     #include "platform/unix_base_memory.c"
     #include "platform/linux_clock.c"
     #include "platform/posix_io.c"
+    #include "platform/platform_io_dialog.c"
     #include "platform/posix_thread.c"
 /*
 	#include"platform/unix_rng.c"
@@ -67,42 +69,47 @@
 #include "util/strings.c"
 #include "util/utf8.c"
 
+#if !defined(OC_NO_APP_LAYER)
 //---------------------------------------------------------------
 // app/graphics layer
 //---------------------------------------------------------------
 
-#if OC_PLATFORM_WINDOWS
-    #include "app/win32_app.c"
-    #include "graphics/win32_vsync.c"
-    #include "graphics/graphics_common.c"
-    #include "graphics/graphics_surface.c"
+    #if OC_PLATFORM_WINDOWS
+        #include "platform/platform_io_dialog.c"
+        #include "app/win32_app.c"
+        #include "graphics/win32_vsync.c"
+        #include "graphics/graphics_common.c"
+        #include "graphics/graphics_surface.c"
 
-    #if OC_COMPILE_GL || OC_COMPILE_GLES
-        #include "graphics/gl_loader.c"
+        #if OC_COMPILE_GL || OC_COMPILE_GLES
+            #include "graphics/gl_loader.c"
+        #endif
+
+        #if OC_COMPILE_GL
+            #include "graphics/wgl_surface.c"
+        #endif
+
+        #if OC_COMPILE_CANVAS
+            #include "graphics/gl_canvas.c"
+        #endif
+
+        #if OC_COMPILE_GLES
+            #include "graphics/egl_surface.c"
+        #endif
+
+    #elif OC_PLATFORM_MACOS
+        #include "platform/platform_io_dialog.c"
+    //NOTE: macos application layer and graphics backends are defined in orca.m
+    #elif OC_PLATFORM_ORCA
+        #include "app/orca_app.c"
+        #include "wasmbind/core_api_stubs.c"
+        #include "graphics/graphics_common.c"
+        #include "graphics/orca_surface_stubs.c"
+    #else
+        #error "Unsupported platform"
     #endif
 
-    #if OC_COMPILE_GL
-        #include "graphics/wgl_surface.c"
-    #endif
+    #include "ui/input_state.c"
+    #include "ui/ui.c"
 
-    #if OC_COMPILE_CANVAS
-        #include "graphics/gl_canvas.c"
-    #endif
-
-    #if OC_COMPILE_GLES
-        #include "graphics/egl_surface.c"
-    #endif
-
-#elif OC_PLATFORM_MACOS
-//NOTE: macos application layer and graphics backends are defined in orca.m
-#elif OC_PLATFORM_ORCA
-    #include "app/orca_app.c"
-    #include "wasmbind/core_api_stubs.c"
-    #include "graphics/graphics_common.c"
-    #include "graphics/orca_surface_stubs.c"
-#else
-    #error "Unsupported platform"
-#endif
-
-#include "ui/input_state.c"
-#include "ui/ui.c"
+#endif // !defined(OC_NO_APP_LAYER)
