+++ src/orca.h
@@ -21,33 +21,41 @@
 #include "platform/platform.h"
 #include "platform/platform_clock.h"
 #include "platform/platform_io.h"
-#include "platform/platform_io_dialog.h"
 #include "platform/platform_path.h"
 
 #if !defined(OC_PLATFORM_ORCA) || !(OC_PLATFORM_ORCA)
     #include "platform/platform_thread.h"
 #endif
 
-#include "app/app.h"
-
-//----------------------------------------------------------------
-// graphics
-//----------------------------------------------------------------
-#include "graphics/graphics.h"
-
-#if OC_PLATFORM_ORCA
-    //TODO: maybe make this conditional
-    #include "graphics/orca_gl31.h"
-
-#else
-    #ifdef OC_INCLUDE_GL_API
-        #include "graphics/gl_api.h"
+#if defined(OC_NO_APP_LAYER)
+    #if OC_PLATFORM_ORCA
+        #error "OC_NO_APP_LAYER can't be used with OC_PLATFORM_ORCA"
     #endif
-#endif
-//----------------------------------------------------------------
-// UI
-//----------------------------------------------------------------
-#include "ui/input_state.h"
-#include "ui/ui.h"
+#else
+
+    #include "app/app.h"
+    #include "platform/platform_io_dialog.h"
+
+    //----------------------------------------------------------------
+    // graphics
+    //----------------------------------------------------------------
+    #include "graphics/graphics.h"
+
+    #if OC_PLATFORM_ORCA
+        //TODO: maybe make this conditional
+        #include "graphics/orca_gl31.h"
+
+    #else
+        #ifdef OC_INCLUDE_GL_API
+            #include "graphics/gl_api.h"
+        #endif
+    #endif
+    //----------------------------------------------------------------
+    // UI
+    //----------------------------------------------------------------
+    #include "ui/input_state.h"
+    #include "ui/ui.h"
+
+#endif // defined(OC_NO_APP_LAYER)
 
 #endif //__ORCA_H_
