+++ src/platform/osx_path.c
@@ -6,7 +6,6 @@
 *
 **************************************************************************/
 
-#import <Foundation/Foundation.h>
 #include <libgen.h>
 #include <mach-o/dyld.h>
 
@@ -19,17 +18,14 @@ bool oc_path_is_absolute(oc_str8 path)
 
 oc_str8 oc_path_executable(oc_arena* arena)
 {
-    @autoreleasepool
-    {
-        oc_str8 result = {};
-        u32 size = 0;
-        _NSGetExecutablePath(0, &size);
-        result.len = size;
-        result.ptr = oc_arena_push_array(arena, char, result.len + 1);
-        _NSGetExecutablePath(result.ptr, &size);
-        result.ptr[result.len] = '\0';
-        return (result);
-    }
+    oc_str8 result = {};
+    u32 size = 0;
+    _NSGetExecutablePath(0, &size);
+    result.len = size;
+    result.ptr = oc_arena_push_array(arena, char, result.len + 1);
+    _NSGetExecutablePath(result.ptr, &size);
+    result.ptr[result.len] = '\0';
+    return (result);
 }
 
 oc_str8 oc_path_canonical(oc_arena* arena, oc_str8 path)
