+++ src/platform/platform_io_common.c
@@ -7,7 +7,6 @@
 **************************************************************************/
 
 #include "platform_io.h"
-#include "platform_io_dialog.h"
 
 //------------------------------------------------------------------------------
 // File stream read/write API
